Replace home page with discover page
