+++ package.json
@@ -30,6 +30,7 @@
     "next-seo": "^4.4.0",
     "react": "16.13.0",
     "react-dom": "16.13.0",
+    "react-xmasonry": "^3.0.3",
     "tailwindcss": "^1.2.0"
   },
   "devDependencies": {
